step: move spaceTimeChart types and colors
Moved WorkSchedule type to types.ts

Moved spaceTimeChart colors to the common colors file
