+++ front/ui/ui-charts/src/spaceTimeChart/lib/consts.ts
@@ -1,25 +1,12 @@
-import { SECOND, MINUTE, HOUR } from '../../common/consts';
+import { SECOND, MINUTE, HOUR, FONT_SIZE, FONT } from '../../common/consts';
+import { BLACK_100, SKY_700, GREY_10, GREY_30, GREY_50 } from '../../common/helpers/colors';
 import { type SpaceTimeChartTheme } from './types';
 
-export const BLACK = '#000000';
-export const BLUE = '#2170B9';
-export const GREY_10 = '#EDEDED';
-export const GREY_30 = '#B6B3AF';
-export const GREY_50 = '#797671';
-export const WHITE = '#FFFFFF';
-export const WHITE_75 = '#FFFFFFC0';
-export const AMBIANT_A10 = '#EFF3F5';
-export const ERROR_30 = '#FF6868';
-export const ERROR_60 = '#D91C1C';
 
 // Occupancy blocks colors :
 export const OCCUPANCY_FREE = '#CAEDDB';
 export const OCCUPANCY_SEMAPHORE = '#FFD4D8';
 export const OCCUPANCY_WARNING = '#FFEABF';
-
-// Fonts:
-export const FONT_SIZE = 10;
-export const FONT = 'IBM Plex Mono, monospace';
 
 export const DEFAULT_THEME: SpaceTimeChartTheme = {
   background: 'white',
@@ -44,12 +31,12 @@ export const DEFAULT_THEME: SpaceTimeChartTheme = {
   spaceGraduationsStyles: {
     1: {
       width: 0.5,
-      color: BLUE,
+      color: SKY_700,
       opacity: 0.75,
     },
     2: {
       width: 0.5,
-      color: BLUE,
+      color: SKY_700,
       opacity: 0.25,
     },
     3: {
@@ -75,24 +62,24 @@ export const DEFAULT_THEME: SpaceTimeChartTheme = {
   timeCaptionsStyles: {
     1: {
       color: GREY_50,
-      font: `12px ${FONT}`,
+      font: `12px ${FONT}, monospace';`,
       topOffset: 12,
     },
     2: {
       color: GREY_30,
-      font: `12px ${FONT}`,
+      font: `12px ${FONT}, monospace';`,
       topOffset: 12,
     },
     3: {
       color: GREY_30,
-      font: `10px ${FONT}`,
+      font: `10px ${FONT}, monospace';`,
       topOffset: 6,
     },
   },
   dateCaptionsSize: 0,
   dateCaptionsStyle: {
     color: GREY_30,
-    font: `10px ${FONT}`,
+    font: `10px ${FONT}, monospace';`,
     topOffset: 28,
     textAlign: 'left',
   },
@@ -113,35 +100,35 @@ export const DEFAULT_THEME: SpaceTimeChartTheme = {
   timeGraduationsStyles: {
     1: {
       width: 0.75,
-      color: BLACK,
+      color: BLACK_100,
       opacity: 0.5,
     },
     2: {
       width: 0.5,
-      color: BLUE,
+      color: SKY_700,
       opacity: 0.7,
     },
     3: {
       width: 0.5,
-      color: BLUE,
+      color: SKY_700,
       opacity: 0.7,
       dashArray: [80, 8],
     },
     4: {
       width: 0.5,
-      color: BLUE,
+      color: SKY_700,
       opacity: 0.7,
       dashArray: [16, 4],
     },
     5: {
       width: 0.5,
-      color: BLUE,
+      color: SKY_700,
       opacity: 0.7,
       dashArray: [8, 8],
     },
     6: {
       width: 0.5,
-      color: BLUE,
+      color: SKY_700,
       opacity: 0.25,
     },
   },
