+++ front/ui/ui-charts/src/spaceTimeChart/lib/types.ts
@@ -155,3 +155,10 @@ export type SpaceTimeChartContextType = TimeChartContextType & {
 
   captionSize: number;
 };
+
+export type WorkSchedule = {
+  type: 'TRACK' | 'CATENARY';
+  timeStart: Date;
+  timeEnd: Date;
+  spaceRanges: [number, number][];
+};
